fix tests again, restructure such that typing actually works
